rename local-apis -> internal/k8s-apis
