+++ Makefile
@@ -63,6 +63,6 @@ release: check-release-deps
 
 .PHONY: update
 update:
-	./local-apis/update.sh
+	./internal/k8s-apis/update.sh
 	go get -u ./...
 	go mod tidy
