+++ internal/k8s-apis/README.md
@@ -1,3 +1,3 @@
-# local-apis
+# k8s-apis
 
 Stats directory is generated from a part of k8s.io/kubernetes.
