+++ internal/k8s-apis/stats/v1alpha1/BUILD
