+++ platform/kubernetes/metric.go
@@ -5,7 +5,7 @@ import (
 	"strconv"
 	"time"
 
-	kubeletTypes "github.com/mackerelio/mackerel-container-agent/local-apis/stats/v1alpha1"
+	kubeletTypes "github.com/mackerelio/mackerel-container-agent/internal/k8s-apis/stats/v1alpha1"
 	kubernetesTypes "k8s.io/api/core/v1"
 
 	mackerel "github.com/mackerelio/mackerel-client-go"
