+++ platform/kubernetes/metric_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 
 	cadvisorTypes "github.com/google/cadvisor/info/v1"
-	kubeletTypes "github.com/mackerelio/mackerel-container-agent/local-apis/stats/v1alpha1"
+	kubeletTypes "github.com/mackerelio/mackerel-container-agent/internal/k8s-apis/stats/v1alpha1"
 	kubernetesTypes "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
